Fix lon/lat var type. Pass module tests

+++ main-service/src/main/java/ru/practicum/main/location/dto/LocationDto.java
@@ -6,6 +6,6 @@ import lombok.*;
 @Setter
 public class LocationDto {
     private Long id;
-    private Double lat;
-    private Double lon;
+    private Float lat;
+    private Float lon;
 }
